backend/requirements.txt
asgiref==3.8.1
astroid==3.3.10
async-timeout==5.0.1
cffi==1.17.1
click==8.2.1
colorama==0.4.6
cryptography==45.0.4
dill==0.4.0
dnspython==2.7.0
ecdsa==0.19.1
fastapi==0.68.2
greenlet==3.2.3
h11==0.16.0
isort==6.0.1
mccabe==0.7.0
passlib==1.7.4
platformdirs==4.3.8
psycopg==3.2.9
pyasn1==0.6.1
pycparser==2.22
pydantic==1.10.22
pylint==3.3.7
pymongo==4.13.2
python-jose==3.5.0
redis==6.2.0
rsa==4.9.1
six==1.17.0
SQLAlchemy==2.0.41
starlette==0.14.2
tomli==2.2.1
tomlkit==0.13.3
typing_extensions==4.14.0
tzdata==2025.2
uvicorn==0.15.0